refactor VBCard for mobile
