+++ template/src/components/precall/LocalMute.tsx
@@ -25,7 +25,6 @@ import LocalAudioMute, {
   LocalAudioMuteProps,
 } from '../../subComponents/LocalAudioMute';
 
-import Spacer from '../../atoms/Spacer';
 import {usePreCall} from './usePreCall';
 import DeviceContext from '../DeviceContext';
 import VBButton from '../virtual-background/VBButton';
@@ -122,32 +121,24 @@ const PreCallLocalMute = (props: PreCallProps) => {
           paddingVertical: 10,
           borderBottomLeftRadius: 8,
           borderBottomRightRadius: 8,
+          flex: 1,
         },
       ]}
       testID="precall-controls">
       <AudioMute showToolTip={true} />
 
       {!$config.AUDIO_ROOM && (
-        <>
-          <Spacer size={isMobileView ? 24 : 16} horizontal={true} />
+        <View style={{marginLeft: isMobileView ? 24 : 16}}>
           <VideoMute showToolTip={true} />
-        </>
+        </View>
       )}
 
-      {/* Settings View in Mobile */}
-      {/* {isMobileView && !isNative && (
-        <>
-          <Spacer size={isMobileView ? 24 : 16} horizontal={true} />
-          <PreCallSettings />
-        </>
-      )} */}
       {$config.ENABLE_VIRTUAL_BACKGROUND &&
         !$config.AUDIO_ROOM &&
         isMobileView && (
-          <>
-            <Spacer size={isMobileView ? 24 : 16} horizontal={true} />
+          <View style={{marginLeft: isMobileView ? 24 : 16}}>
             <VBButton isVBOpen={isVBOpen} setIsVBOpen={setIsVBOpen} />
-          </>
+          </View>
         )}
     </View>
   );
@@ -191,7 +182,5 @@ const style = StyleSheet.create({
     justifyContent: 'center',
     alignItems: 'center',
     backgroundColor: $config.CARD_LAYER_1_COLOR,
-    // borderBottomLeftRadius: 4,
-    // borderBottomRightRadius: 4,
   },
 });
