+++ template/src/components/virtual-background/VBCard.native.tsx
@@ -1,0 +1,14 @@
+import {StyleSheet, Text, View} from 'react-native';
+import React from 'react';
+
+const VBCard = () => {
+  return (
+    <View>
+      <Text>VBCard.native</Text>
+    </View>
+  );
+};
+
+export default VBCard;
+
+const styles = StyleSheet.create({});
