+++ template/src/components/virtual-background/VideoPreview.native.tsx
@@ -1,0 +1,14 @@
+import {StyleSheet, Text, View} from 'react-native';
+import React from 'react';
+
+const VideoPreview = () => {
+  return (
+    <View>
+      <Text>VideoPreview.native</Text>
+    </View>
+  );
+};
+
+export default VideoPreview;
+
+const styles = StyleSheet.create({});
